@@ -2,37 +2,36 @@
   "name": "react-template",
   "private": true,
   "dependencies": {
-    "@types/react-ga": "^2.3.0",
     "buffer": "^6.0.3",
-    "html-react-parser": "^1.4.8",
+    "html-react-parser": "^1.4.11",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-fetching-library": "^1.7.6",
-    "react-ga": "^3.3.0",
-    "react-redux": "^7.2.6",
-    "react-router": "^6.2.1",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "^5.0.0",
+    "react-ga4": "^1.4.1",
+    "react-redux": "^8.0.0",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
     "react-toastify": "^8.2.0",
     "redux": "^4.1.2",
-    "redux-thunk": "^2.4.1",
-    "sass": "^1.49.7"
+    "sass": "^1.50.0",
+    "web-vitals": "^2.1.4"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.178",
-    "@types/node": "^16.11.25",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-redux": "^7.1.22",
+    "@types/lodash": "^4.14.181",
+    "@types/node": "^17.0.24",
+    "@types/react-dom": "^18.0.1",
+    "@types/react-redux": "^7.1.24",
     "@types/react-router": "^5.1.18",
     "@types/react-router-dom": "^5.3.3",
-    "@types/webpack-env": "^1.16.3",
-    "@typescript-eslint/eslint-plugin": "^5.12.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "eslint-plugin-react": "^7.28.0",
+    "@types/webpack-env": "^1.16.4",
+    "@typescript-eslint/eslint-plugin": "^5.19.0",
+    "@typescript-eslint/parser": "^5.19.0",
+    "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint-plugin-unused-imports": "^2.0.0",
-    "typescript": "^4.5.5"
+    "typescript": "^4.6.3"
   },
   "scripts": {
     "start": "react-scripts start",
