@@ -1,8 +1,12 @@
 import App from "app/App"
 import { createElement } from "react"
-import ReactDOM from "react-dom"
+import { createRoot } from "react-dom/client"
 
 require("polyfills")
 require("preload")
 
-ReactDOM.render(createElement(App), document.getElementById("root"))
+const rootElement = document.getElementById("root")
+if (rootElement) {
+  const root = createRoot(rootElement)
+  root.render(createElement(App))
+}
