@@ -1,7 +1,7 @@
 import "./Button.scss"
 
 import { MouseEvent, MouseEventHandler, useState } from "react"
-import ReactGA from "react-ga"
+import ReactGA from "react-ga4"
 import { classMerge, classWithModifiers } from "utils/common"
 
 import Loader from "../Loader/Loader"
