@@ -1,11 +1,7 @@
-import "react-redux"
-
-import { applyMiddleware, compose, createStore } from "redux"
-import thunk from "redux-thunk"
+import { createStore } from "redux"
 
 import combinedReducers from "./combinedReducers"
 
-const enhancer = compose(applyMiddleware(thunk))
-const store = createStore(combinedReducers, enhancer)
+const store = createStore(combinedReducers)
 
 export default store
